rem silent pulling
in my opinion i wanted to suppress warn messages that are not correct in this step. that i can set the creditials i must befor pull the images because i must generate the password hash with the hornet tool (e.g.). but maybe there is an other solution. we will remove this from this PR and give it on our roadmap