improved version with saved peers
This improved version offers a more intuitive approach to managing auto-peering and configuring static neighbors. To activate auto peering, users can simply press "E" and Enter. To disable auto peering, users can enter static neighbors by pressing the "X" key. The configurations are saved in the .env file using the format "Alias:MultiAddress, ...".

When running prepare_docker.sh, the script extracts the configuration data from the .env file and converts it to the peering.json format and also handles spaces and the absence of an alias. This file is then stored in the specified folder to ensure smooth integration into the networks.